Switch to a model-centric approach
This adds a model for each of the "screens".

Note that each model's Update function is currently a duplicated
rendering of a menu. I'm ok with that for now, as it let's me easily
navigate through each model. Obviously, each function will change, and I
may choose to abstract away the menu and its interface later if I deem
it necessary.

This also renames Init to Reset. This is due to the way Golang handles
the token "init".